UT:Zero iterations REP-string test
When a REP-string is executed in 64-bit mode with an address-size prefix,
ECX/EDI/ESI are used as counter and pointers. When ECX is initially zero, Intel
CPUs clear the high 32-bits of RCX, and recent Intel CPUs update the high bits
of the pointers in MOVS/STOS.

This is an undocumented behavior.

Signed-off-by: Xiangyang Wu <xiangyang.wu@linux.intel.com>